Fix 64-bit decompression bugs and a data race across PPMD, Dict, GRZip, and TTA

+++ Compression/MM/entropy.h
@@ -41,10 +41,10 @@ extern unsigned char *bit_array_read;
 extern unsigned char *bit_array_write;
 
 void init_bit_array_write (void);
-void init_bit_array_read (unsigned long size);
-long get_len (void);
+void init_bit_array_read (uint32 size);
+int32 get_len (void);
 
-void encode_frame (long *data, unsigned long len);
-void decode_frame (long *data, unsigned long len);
+void encode_frame (int32 *data, uint32 len);
+void decode_frame (int32 *data, uint32 len);
 
 #endif  /* ENTROPY_H */
